requirements.txt
flask==3.0.3
chromadb==0.4.24
pypdf==4.2.0
google-generativeai==0.4.0
python-dotenv==1.0.1
gunicorn==22.0.0
numpy<2.0

